Fixed from mistakes, rearranged the files and renamed a couple things. Updated all file paths accordingly.
