+++ client/src/App.js
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
 import './App.css';
-import './components/create/create.css';
+import './components/create_user/create.css';
 import './components/Navbar/Navbar.css';
 import Login from './components/login/login.js';
 import CamiBookNav from './components/Navbar/Navbar.js';
+import Feed from './components/feed/feed.js';
 
 
 
@@ -13,6 +14,7 @@ class App extends Component {
       <div>
         <h1 className="banner">Welcome to our App</h1>
           <CamiBookNav />
+          <Feed />
           <Login />
       </div>
     );
